team-98-master 2/requirements.txt
mysql-connector-python
pandas
pymysql
pymysql
calmap
matplotlib
numpy
